Move UseMeaningfulNameInspectionTests for QuickFixes into relevant files under RD.QuickFixes
