+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -1000,6 +1000,38 @@ End Sub";
             Assert.AreEqual(expectedCode, parser.State.GetRewriter(component).GetText());
         }
 
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void UseMeaningfulName_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Ffffff()
+End Sub";
+
+            const string expectedCode =
+@"'@Ignore UseMeaningfulName
+Sub Ffffff()
+End Sub";
+
+            var builder = new MockVbeBuilder();
+            var project = builder.ProjectBuilder("VBAProject", ProjectProtection.Unprotected)
+                .AddComponent("MyClass", ComponentType.ClassModule, inputCode)
+                .Build();
+            var component = project.Object.VBComponents[0];
+            var vbe = builder.AddProject(project).Build();
+
+            var parser = MockParser.Create(vbe.Object);
+
+            parser.Parse(new CancellationTokenSource());
+            if (parser.State.Status >= ParserState.Error) { Assert.Inconclusive("Parser Error"); }
+
+            var inspection = new UseMeaningfulNameInspection(parser.State, UseMeaningfulNameInspectionTests.GetInspectionSettings().Object);
+            var inspectionResults = inspection.GetInspectionResults();
+
+            new IgnoreOnceQuickFix(parser.State, new[] { inspection }).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, parser.State.GetRewriter(component).GetText());
+        }
+
 
 
     }
